Revert "[feature](load) enable memtable on sink node (#23074)"
This reverts commit f747cc8a9429fd3454c16f3a3523ee466f2aa64d.
