+++ be/src/io/fs/file_writer.h
@@ -32,7 +32,6 @@ namespace io {
 class FileWriter {
 public:
     FileWriter(Path&& path, FileSystemSPtr fs) : _path(std::move(path)), _fs(fs) {}
-    FileWriter() = default;
     virtual ~FileWriter() = default;
 
     DISALLOW_COPY_AND_ASSIGN(FileWriter);
@@ -58,8 +57,6 @@ public:
     size_t bytes_appended() const { return _bytes_appended; }
 
     FileSystemSPtr fs() const { return _fs; }
-
-    bool is_closed() { return _closed; }
 
 protected:
     Path _path;
