+++ be/src/olap/delta_writer_context.h
@@ -27,7 +27,6 @@ namespace doris {
 class TupleDescriptor;
 class SlotDescriptor;
 class OlapTableSchemaParam;
-class TabletSchema;
 
 struct WriteRequest {
     int64_t tablet_id;
@@ -41,10 +40,6 @@ struct WriteRequest {
     bool is_high_priority = false;
     OlapTableSchemaParam* table_schema_param;
     int64_t index_id = 0;
-    // for DeltaWriterV2
-    std::shared_ptr<TabletSchema> tablet_schema;
-    bool enable_unique_key_merge_on_write = false;
-    int sender_id;
 };
 
 } // namespace doris
