+++ be/src/olap/rowset/beta_rowset_writer.h
@@ -62,6 +62,13 @@ namespace vectorized::schema_util {
 class LocalSchemaChangeRecorder;
 }
 
+struct SegmentStatistics {
+    int64_t row_num;
+    int64_t data_size;
+    int64_t index_size;
+    KeyBoundsPB key_bounds;
+};
+
 class BetaRowsetWriter : public RowsetWriter {
     friend class SegcompactionWorker;
 
@@ -79,9 +86,9 @@ public:
 
     Status add_rowset_for_linked_schema_change(RowsetSharedPtr rowset) override;
 
-    Status create_file_writer(uint32_t segment_id, io::FileWriterPtr& writer) override;
+    Status create_file_writer(uint32_t segment_id, io::FileWriterPtr& writer);
 
-    Status add_segment(uint32_t segid, SegmentStatistics& segstat) override;
+    Status add_segment(uint32_t segid, SegmentStatistics& segstat);
 
     Status flush() override;
