+++ be/src/olap/rowset/rowset_writer.h
@@ -24,7 +24,6 @@
 #include <optional>
 
 #include "common/factory_creator.h"
-#include "gen_cpp/internal_service.pb.h"
 #include "gutil/macros.h"
 #include "olap/column_mapping.h"
 #include "olap/rowset/rowset.h"
@@ -34,35 +33,7 @@
 
 namespace doris {
 
-struct SegmentStatistics {
-    int64_t row_num;
-    int64_t data_size;
-    int64_t index_size;
-    KeyBoundsPB key_bounds;
-
-    SegmentStatistics() = default;
-
-    SegmentStatistics(SegmentStatisticsPB pb)
-            : row_num(pb.row_num()),
-              data_size(pb.data_size()),
-              index_size(pb.index_size()),
-              key_bounds(pb.key_bounds()) {}
-
-    void to_pb(SegmentStatisticsPB* segstat_pb) {
-        segstat_pb->set_row_num(row_num);
-        segstat_pb->set_data_size(data_size);
-        segstat_pb->set_index_size(index_size);
-        segstat_pb->mutable_key_bounds()->CopyFrom(key_bounds);
-    }
-
-    std::string to_string() {
-        std::stringstream ss;
-        ss << "row_num: " << row_num << ", data_size: " << data_size
-           << ", index_size: " << index_size << ", key_bounds: " << key_bounds.ShortDebugString();
-        return ss.str();
-    }
-};
-using SegmentStatisticsSharedPtr = std::shared_ptr<SegmentStatistics>;
+class MemTable;
 
 class RowsetWriter {
 public:
@@ -87,11 +58,6 @@ public:
     // Precondition: the input `rowset` should have the same type of the rowset we're building
     virtual Status add_rowset_for_linked_schema_change(RowsetSharedPtr rowset) = 0;
 
-    virtual Status create_file_writer(uint32_t segment_id, io::FileWriterPtr& writer) {
-        return Status::Error<ErrorCode::NOT_IMPLEMENTED_ERROR>(
-                "RowsetWriter not support create_file_writer");
-    }
-
     // explicit flush all buffered rows into segment file.
     // note that `add_row` could also trigger flush when certain conditions are met
     virtual Status flush() = 0;
@@ -114,8 +80,6 @@ public:
         return Status::Error<ErrorCode::NOT_IMPLEMENTED_ERROR>(
                 "RowsetWriter not support flush_single_block");
     }
-
-    virtual Status add_segment(uint32_t segid, SegmentStatistics& segstat) = 0;
 
     // finish building and return pointer to the built rowset (guaranteed to be inited).
     // return nullptr when failed
