+++ be/src/olap/rowset/segment_v2/segment_writer.cpp
@@ -37,7 +37,6 @@
 #include "olap/olap_common.h"
 #include "olap/primary_key_index.h"
 #include "olap/row_cursor.h"                      // RowCursor // IWYU pragma: keep
-#include "olap/rowset/rowset_writer.h"            // SegmentStatistics
 #include "olap/rowset/rowset_writer_context.h"    // RowsetWriterContext
 #include "olap/rowset/segment_v2/column_writer.h" // ColumnWriter
 #include "olap/rowset/segment_v2/page_io.h"
