+++ fe/fe-core/src/main/java/org/apache/doris/task/LoadTaskInfo.java
@@ -121,10 +121,6 @@ public interface LoadTaskInfo {
         return false;
     }
 
-    default boolean isMemtableOnSinkNode() {
-        return false;
-    }
-
     class ImportColumnDescs {
         public List<ImportColumnDesc> descs = Lists.newArrayList();
         public boolean isColumnDescsRewrited = false;
