+++ fe/fe-core/src/main/java/org/apache/doris/task/StreamLoadTask.java
@@ -88,8 +88,6 @@ public class StreamLoadTask implements LoadTaskInfo {
     private int skipLines = 0;
     private boolean enableProfile = false;
 
-    private boolean memtableOnSinkNode = false;
-
     private byte enclose = 0;
 
     private byte escape = 0;
@@ -299,15 +297,6 @@ public class StreamLoadTask implements LoadTaskInfo {
         return isPartialUpdate;
     }
 
-    @Override
-    public boolean isMemtableOnSinkNode() {
-        return memtableOnSinkNode;
-    }
-
-    public void setMemtableOnSinkNode(boolean memtableOnSinkNode) {
-        this.memtableOnSinkNode = memtableOnSinkNode;
-    }
-
     public static StreamLoadTask fromTStreamLoadPutRequest(TStreamLoadPutRequest request) throws UserException {
         StreamLoadTask streamLoadTask = new StreamLoadTask(request.getLoadId(), request.getTxnId(),
                 request.getFileType(), request.getFormatType(),
@@ -440,9 +429,6 @@ public class StreamLoadTask implements LoadTaskInfo {
         }
         if (request.isSetPartialUpdate()) {
             isPartialUpdate = request.isPartialUpdate();
-        }
-        if (request.isSetMemtableOnSinkNode()) {
-            this.memtableOnSinkNode = request.isMemtableOnSinkNode();
         }
     }
