+++ gensrc/thrift/FrontendService.thrift
@@ -614,7 +614,6 @@ struct TStreamLoadPutRequest {
     51: optional i8 enclose
     // only valid when file type is CSV
     52: optional i8 escape
-    53: optional bool memtable_on_sink_node;
 }
 
 struct TStreamLoadPutResult {
