+++ gensrc/thrift/PaloInternalService.thrift
@@ -231,8 +231,6 @@ struct TQueryOptions {
   77: optional bool truncate_char_or_varchar_columns = false
 
   78: optional bool enable_hash_join_early_start_probe = false
-
-  79: optional bool enable_memtable_on_sink_node = false;
 }
